fix bugs

+++ app/src/main/java/com/tasdjilati/ui/main/enter/StudentsEnterFragment.kt
@@ -1,10 +1,12 @@
 package com.tasdjilati.ui.main.enter
 
 import android.Manifest
+import android.app.Activity
+import android.app.AlertDialog
 import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
+import android.content.*
 import android.content.pm.PackageManager
+import android.net.Uri
 import android.os.Bundle
 import android.telephony.SmsManager
 import android.view.LayoutInflater
@@ -15,12 +17,14 @@ import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.tasdjilati.data.entities.Student
 import com.tasdjilati.data.entities.StudentEnterAttendance
 import com.tasdjilati.databinding.FragmentStudentsEnterBinding
+import com.tasdjilati.ui.main.exit.StudentExitAttendanceAdapter
 import com.tasdjilati.ui.main.students_list.StudentViewModel
-import java.lang.Exception
+import kotlinx.coroutines.*
 
 class StudentsEnterFragment : Fragment() {
     private val REQUEST_CODE: Int = 1
@@ -46,12 +50,62 @@ class StudentsEnterFragment : Fragment() {
 
 
         binding.btnCancel.setOnClickListener {
-            deleteStudentAttendanceList()
+            val builder: AlertDialog.Builder = context.let {
+                AlertDialog.Builder(it)
+            }
+
+            builder.setMessage("Voulez vous vraiment annuler le processus ?")
+                .setTitle("Annuler")
+
+            builder.apply {
+                setPositiveButton("Oui") { dialog, id ->
+                    try{
+                        deleteStudentAttendanceList()
+                        insertStudents()
+
+                    }catch(e : Exception){
+
+                    }
+                }
+                setNegativeButton("Non") { dialog, id ->
+                    dialog.dismiss()
+                }
+            }
+            val dialog: AlertDialog? = builder.create()
+
+            dialog!!.show()
         }
 
         binding.btnTerminate.setOnClickListener {
-            sendSmsToAbsents()
-            deleteStudentAttendanceList()
+            try{
+                val builder: AlertDialog.Builder = context.let {
+                    AlertDialog.Builder(it)
+                }
+
+                builder.setMessage("Voulez vous vraiment terminer le processus ?")
+                    .setTitle("Terminer")
+
+                builder.apply {
+                    setPositiveButton("Oui") { dialog, id ->
+                        try{
+                            sendSmsToAbsents()
+                            deleteStudentAttendanceList()
+                            insertStudents()
+
+                        }catch(e : Exception){
+
+                        }
+                    }
+                    setNegativeButton("Non") { dialog, id ->
+                        dialog.dismiss()
+                    }
+                }
+                val dialog: AlertDialog? = builder.create()
+
+                dialog!!.show()
+            }catch (e : Exception){
+
+            }
         }
 
         return binding.root
@@ -63,24 +117,33 @@ class StudentsEnterFragment : Fragment() {
 
     }
 
-    private fun sendSmsToAbsents() {
-        for (student in studentsList!!){
-            if (student.attendance == 0)
-                sendSMS(student , "Votre enfant est absent")
+    private fun sendSmsToAbsents()= lifecycleScope.launch{
+        for (student in studentsList!!)  {
+            if (student.attendance == 0){
+                sendSMS(student.numParent1, "Votre enfant ${student.name} ${student.surname} est absent, veuillez justifier la cause")
+                delay(200L)
+                sendSMS(student.numParent2 , "Votre enfant ${student.name} ${student.surname} est absent, veuillez justifier la cause")
+                delay(200L)
+            }
+
         }
     }
 
-    private fun sendSMS(student : StudentEnterAttendance, message : String) {
-        if (checkSmsPermission()){
-                try {
-                    val sentPI: PendingIntent = PendingIntent.getBroadcast(requireContext(), 0, Intent("SMS_SENT"), 0)
-                    SmsManager.getDefault().sendTextMessage(student.numParent1, null, message, sentPI, null)
-                    SmsManager.getDefault().sendTextMessage(student.numParent2, null, message, sentPI, null)
-                }catch (e : Exception){
+    private fun sendSMS(phoneNumber: String, message: String) {
+        if(checkSmsPermission() && phoneNumber.isNotEmpty()){
+            val SENT = "SMS_SENT"
+            val DELIVERED = "SMS_DELIVERED"
+            val sentPI = PendingIntent.getBroadcast(activity!!, 0, Intent(
+                SENT), 0)
+            val deliveredPI = PendingIntent.getBroadcast(activity!!, 0,
+                Intent(DELIVERED), 0)
+            val sms = SmsManager.getDefault()
+            sms.sendTextMessage(phoneNumber, null, message, sentPI, deliveredPI)
 
-                }
-            }
+
+        }
     }
+
 
     private fun checkSmsPermission() : Boolean{
         return if (ContextCompat.checkSelfPermission(activity!!, Manifest.permission.SEND_SMS)
@@ -106,12 +169,18 @@ class StudentsEnterFragment : Fragment() {
     }
 
     private fun insertStudents() {
-        studentViewModel.getAllStudents.observe(viewLifecycleOwner , { studentsList ->
-            for (student in studentsList){
-                val studentAttendance = StudentEnterAttendance(student.id , student.name , student.surname , student.birthDate ,
+        var studentAttendanceList : ArrayList<StudentEnterAttendance> = java.util.ArrayList()
+        studentViewModel.getAllStudents.observe(viewLifecycleOwner , {
+            for (student in it!!){
+                val studentAttendance = StudentEnterAttendance( student.id , student.name , student.surname , student.birthDate ,
                     student.year , student.classe , student.numParent1 , student.numParent2 , student.address , 0)
+                studentAttendanceList.add(studentAttendance)
+            }
+
+            for (studentAttendance in studentAttendanceList){
                 studentEnterViewModel.addStudent(studentAttendance)
             }
+
         })
 
         val sp = activity?.getSharedPreferences("sp" , Context.MODE_PRIVATE)
